Files for running prototype of game on Unity
These should be all the files necessary? I'm not 100% sure that they will be, though, because I don't know if your libraries contain the same things as mine, or which parts of the contents of my "First Version" folder are necessary parts of the project and which are parts of Unity. If it doesn't work, we can try uploading different things.