Movement While Facing Directions!
This is the AIMovement script modified to use the direction indicator I added, which is a small white triangle that's a child of the Guard object. Visually, it's attached to the "front" of the Guard circle, and it points in the direction the Guard is facing.

Having the guard rotate every 60 frames, as opposed to simply changing direction, makes movement a lot simpler. transform.Translate adds the vector in relation to the direction the guard is facing, so the movement vector never needs to be reassigned anymore. To account for this, I've moved the assignment for that variable into the Start method so that it's not inefficiently being reassigned to a brand-new vector every frame.