Guard & DirectionIndicator Sprites
Not sure if this will work, but I'm trying to upload the files for the objects I'm using for the Guard and DirectionIndicator sprites, so that you guys can just copy them into your versions and not have to mess with the settings in order to create a Guard with a DirectionIndicator. Let me know if this works.