In-Progress Pause Script
The "Pause" script, a component of the PauseHandler object, now includes
code to disable both player and guard movement (including when there are
multiple guards) when the player picks up a pickup. THIS CODE DOES NOT
WORK, but it shouldn't affect anything else we're working on. If it
does, just disable the PauseHandler object, since it doesn't do anything
else.

(The reason this code doesn't work, as far as I can tell, is that the
PauseHandler object has no way of determining when the Player object has
picked up a pickup, or is within the Collider2D of an object tagged
"Pickup". I'm trying to fix this by putting this code in a new public
method, which can be called from the Movement script when an object is
picked up; the problem there is getting the Player object to identify
the PauseHandler object and "know" that this method can be called.
Again, if you guys don't want to mess with this, you don't have to; just
disable the PauseHandler in the hierarchy. These notes are mostly for
myself tomorrow morning.)