More In-Progress Pause Code
Found a way to call the PauseGame method from outside the Pause script;
everything is now connected the way it should be. However, the
"script.enabled = false" command does not seem to disable scripts the
way that Unity tutorials, documentation, and online questions say it
should. This is code I can work from later, but right now it doesn't
actually do anything in-game.