Clean the useless code + let the player stop when being caught
*Clean/comment out the useless code
* let the player stop moving when it is being caught by the guard
instead of disappearing
* some code of chasing the player

There is still a bug in the chasePlayer method, so I currently comment
all the code about this method out. Therefore, now the guard will stop
moving when the player hit the vision cone, but the player can still
move. This bug will be fixed when I uncomment the code about chasing
the player.