Guard Movement Updates
I think the reason the guard was avoiding the player when trying to
chase vertically was because the guard's rotation was throwing off the
direction of its movement vectors, so this code rotates the movement
vector if the guard is pointing up or down. It still doesn't really
work, but it's better than what we had before; it at least moves in the
general direction of the player. Sometimes the guard moves through
walls, though, which is its own problem.