Add Homepage to the game
What’s the name of our game???

I add a homepage to the game. The play button will lead to the first
level we have. The “how to play” button will lead to a instruction,
which I haven’t created yet. And “quit” button will quit the game,
which also does not work now because this is not an actually game that
can be played on the computer yet. (And now the Homepage button when
you pause the game should also work)

I will rearrange the layout and many other visual things after I create
other scenes so that all the settings can be the same.