Instruction page
simple structure of instruction page.

I haven’t had text yet, so only set the instruction page and buttons.

Now play button in the homepage should lead to level select page, which
further lead to the game.
How to play button in the homepage should lead to the instruction,
which will finally also lead to the game(I’m not so sure about this
part…I might change this).
Quit button will quit the game