Third Level First Draft
The idea for this one is that you have to "distract" the guards so that
they'll leave part of the room unguarded long enough that you can grab
the art and get out. Current problems:
- doesn't have doors yet
- the amount of guard movement calculation slows my computer waaaaay
down, so I can't playtest it really at all
- art is still in the default places
- middle room doesn't really do anything--maybe later it should be a
place you can get away from the guards, if the level is too hard?
Probably there should be more art in the middle room, or it wouldn't
make sense to have it there