First Intro Level
Created the first intro level, with just the player, the art, and a
door. BUT the player can move through walls for some reason (only in
this level, not in any of the others).

Also made the camera in Level Three follow the player, which I had
forgotten to do before.