Changed Doors in Levels 1-3
In the previous commit I made a new "Door" prefab, which included a
script to change colors when the player had collected all the art
(rather than having the player's Movement script do it). I also changed
the doors in the intro levels and Level Four to use this prefab.

In this commit, I've changed the doors in Levels 1-3 to use the prefab.
I had to break some Room prefabs to do this. I also reordered some
hierarchies.