In-Progress Meme Variation Code
Made a new UpdateMemes script and attached it to a new prefab,
MemeHolder. Currently only Level One has a MemeHolder, but adding it to
other levels shouldn't be hard; just drag it into the hierarchy and
assign its spriteToPass variable. It doesn't need to be physically
placed into the part of the level that will be visible to the player.

The UpdateMemes script holds a public static variable called
mostRecentSprite and a public (non-static) variable called spriteToPass.
mostRecentSprite is a piece of state shared by the entire game; this is
how we'll tell the transition screen what memes to display. spriteToPass
is specific to the instance of the MemeHolder object, and therefore each
level will have a different spriteToPass. At the beginning of each
level, that level's UpdateMemes script will reassign mostRecentSprite to
be that level's spriteToPass so that, at the end of each level, the
mostRecentSprite that the transition screen accesses will be the
spriteToPass for the level that was just completed.

What I haven't done yet is:
- add instances of the MemeHolder prefab to all the levels besides Intro
Level One
- create Sprite objects to assign to the mostRecentSprite variable
(could also change the type from Sprite to Image, since I don't know how
to create custom sprites, but I assume it's possible so this is your
choice)
- modify the script for the transition screen so that it accesses the
mostRecentSprite variable and displays an image determined by it